Always specify `debootstrap` option `--merged-usr`
Rationale:

Although a straight reading of `debootstrap(8)` suggests that
`--merged-usr` is the default for all variants except `buildd`, this is
not actually the case, at least not for version `1.0.118ubuntu1.3`.
Since buildroot support for the systemd init framework requires a
merged-usr tree, it is more robust to always specify the option. That
way we can support the case where both the buildroot target tree and the
rootfs overlay provide the systemd init framework.