During trip test, use `sudo` when checking for `debootstrap` in PATH
Motivation:

We want to support running the trip test as a non-root user, but a
typical non-root user will not have `/usr/sbin` in his/her PATH, and yet
that is where `debootstrap` is normally located.

Since `debootstrap` is only needed when managing rootfs overlays, and
`sudo(8)` permission is already needed for other rootfs overlay
operations, it is more robust and more general to use root's PATH for
the command check, rather than the PATH of the non-root user running the
trip test.