Add support for inferring and/or skipping particular output trees
An xctc and/or main defconfig will be inferred correctly if the
corresponding type of output tree already exists.

If an xctc and/or main defconfig is not specified and cannot be
inferred, the corresponding output tree(s) are not acted upon.

Note that by design an rol build uses the main buildroot config.