Add script symlinks in `bin` as source code
Motivating use case:

When the trip test performs a main build, `buildroot.sh` invokes `make`,
which eventually invokes the `post-fakeroot` script, which in turns invokes
`buildroot` in order to execute the rootfs overlay post-fakeroot hook. Note
the distinction between `buildroot.sh` and `buildroot`.

Now consider the case when `buildroot-scripts` is *not* installed on the
build host. In that case, the trip test will fail because `buildroot` is
not available, even though `buildroot.sh` is in PATH during the trip
test, by design.

To resolve the matter, plant sibling symlinks alongside each `.sh` file
in the `bin` directory of the source tree. That way `buildroot` found in
PATH whenever `buildroot.sh` is found.